tests: split cli/tests to separate crate, remove test-fakes feature (PoC)
The goal is to remove the crate-cycle: `jj-cli --(dev)--> jj-cli`. Since there's
no way to express binary dependency in Cargo.toml, all binaries which the tests
depend on should have to be built within the same crate.

I put Cargo.toml in cli/tests so the directory structure looks familiar.

One big caveat: since we build a separate "jj" binary for tests, the production
"jj" binary is no longer covered by these tests. We might want to leave a couple
of tests in jj-cli to ensure that the real "jj" binary works.