github: don't run build workflows twice on push to `main`
The new `merge_group` event will already attach all the CI events from the
various workflows where it applies. If the merge queue is drained and merges to
`main`, having `push` on these workflows will cause them to be run again, doubling
the number of CI checks. We don't need `push` anymore, basically.

Also, because checks can be cancelled now, the current double running can
probably lead to an awkward setup like:

- Merge Queue merges A to main, starts `push` workflows
- Merge Queue then merges B and C to main, starts `push` workflows
- `A`'s workflows get cancelled if it's not yet done
- This makes it look like `A` has failing tests somehow, but it doesn't

Therefore, just removing the double builds is the first step to cut down on the
CI times for our repo.

We also run the build workflows on all pushes to NOT main, because that will
help people who want to run CI before opening an actual PR.

Signed-off-by: Austin Seipp <aseipp@pobox.com>