workflows: fix the pre-release docs workflow
It contained a check for `repo_owner == 'martinvonz'` which meant that it hasn't run in 4 weeks.
This is another great example showing why current CI systems suck, as it always was "successful" in
doing nothing with no warning whatsoever.

I also only noticed it because I wanted to check my latest doc change on the prerelease site.