index: don't preserve commits referenced as predecessors
I originally considered it would be a bad idea to ignore predecessors that don't
exist in the store because doing that could shadow potential data corruption
issues. However, we have index built from the op log, which knows all reachable
commits. So we can ask index whether the predecessor id is supposed to exist.
If a predecessor exists in the index but not in the Git object store, the
repository is corrupted.

https://github.com/jj-vcs/jj/pull/4953#issuecomment-2495735683

In this implementation, we assume that predecessor commits are reachable from at
least one of the historical views. Some commands (e.g. squash into descendants)
may create transitive commits within an operation. In that case, the predecessor
chain is initially preserved, but will be lost after reindexing. We might have
to fix these commands to not record transitive predecessors.

Note that this change should be considered forward-incompatible change. We can't
easily revert this patch because the stored commits will have unreachable
predecessors once we run "jj op abandon && jj util gc".