merged_tree: add async version of `path_value()`, use in annotate
`jj file annotate` is very slow at Google. I think it's mostly because
when we walk the tree to find the right FileId, we do so one commit at
a time. We should be able to do those walks concurrently. This patch
is one little step in that direction.

I actually wrote most of this patch for something in
`MergedTreeBuilder`, IIUC, so it should be useful there too.