index: omit operations that should be covered by the parent index
I'm not going to add operation index anytime soon, but this might help if we
decide to. Suppose we index "commit_id: originating_op_id" relations, we'll
probably want to store op_id in a separate sstable to save disk space. If the
table contains all operation ids, we can reuse it to resolve short operation
ids.