cli: improve hint to abandon commits when divergent change appears
A divergent change could point to more than two commits, so the hint now suggests
abandoning unneeded commits rather than just one of them.

The `-r` flag is unnecessary for `jj abandon` as it is implied.

Also changed `<REVISION>` to `<commit_id>` to be clear that using the change id here
will not work.